Mappings for all the apps is done properly.
