+++ src/components/Main.tsx
@@ -2,10 +2,10 @@ import React from 'react';
 import { Route, Switch } from 'react-router-dom';
 import { makeStyles } from '@material-ui/core/styles';
 // import Apps from '../pages/Apps';
-// import Settings from '../pages/Settings';
+import Settings from '../pages/Settings';
 // import UserProfileMappings from '../pages/ProfileMappings';
-import ProfileMappings from '../pages/ProfileMappings2';
-import UserProfileMappings from '../pages/ProfileMappings';
+import ProfileMappings from '../pages/ProfileMappings';
+import InitialNodes from '../pages/InitialNodes';
 
 const useStyles = makeStyles(() => ({
   root: {
@@ -20,10 +20,10 @@ const Main = () => {
   return (
     <div className={classes.root}>
       <Switch>
-        <Route exact path="/" component={UserProfileMappings} />
+        <Route exact path="/" component={InitialNodes} />
         {/* <Route exact path="/" component={Apps} /> */}
         {/* <Route exact path="/apps" component={Apps} /> */}
-        {/* <Route exact path="/settings" component={Settings} /> */}
+        <Route exact path="/settings" component={Settings} />
       </Switch>
       <Switch>
         <Route exact path="/mappings/:id/:label" component={ProfileMappings} />
